Changing pos from int to long

+++ src/main/java/com/julienvey/trello/domain/Card.java
@@ -27,7 +27,7 @@ public class Card extends TrelloEntity {
     private String idShort;
     private String idAttachmentCover;
     private boolean manualCoverAttachment;
-    private int pos;
+    private long pos;
     private String shortLink;
     private String shortUrl;
     private boolean subscribed;
@@ -202,11 +202,11 @@ public class Card extends TrelloEntity {
         this.manualCoverAttachment = manualCoverAttachment;
     }
 
-    public int getPos() {
+    public long getPos() {
         return pos;
     }
 
-    public void setPos(int pos) {
+    public void setPos(long pos) {
         this.pos = pos;
     }
